Bug 2194  	build smoldyn in Visual Studio
svn path=/vcell/trunk/numerics/; revision=8263


Former-commit-id: ceb52054e73585eef22585c5c814b9f00fb084a5

+++ smoldyn-2.15/source/Smoldyn/smolcmd.c
@@ -1956,7 +1956,7 @@ enum CMDcode cmdexcludesphere(simptr sim,cmdptr cmd,char *line2) {
 	itct=sscanf(line2,"%lg",&rad);
 	SCMDCHECK(itct==1,"failure reading radius");
 
-	dist=rad*sqrt(dim);
+	dist=rad*sqrt((double)dim);
 	for(d=0;d<dim;d++) {
 		poslo[d]=poscent[d]-dist;
 		poshi[d]=poscent[d]+dist; }
